moved auth container and changing signin/signout process
