+++ src/index.js
@@ -2,12 +2,11 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './index.css';
-import App from './App';
 import registerServiceWorker from './registerServiceWorker';
 import firebase from 'firebase/app';
 import 'firebase/auth';
-import {AuthContainer} from './SignUp'
 import 'firebase/database';
+import AuthContainer from'./AuthContainer'
 
 var config = {
     apiKey: "AIzaSyAYQnow1D-GsGFX3hsHJ1fKrEPNaQB-D5o",
